+++ ReleaseNotes/release_notes.md
@@ -1,5 +1,7 @@
 ## New features
     
+  1. Add  `cuts` and `cut_range` argumenst for `ds2numpy` function
+        
 ## Backward incompatible
     
 ## Bug fixes